requirements.txt
networkx>=3.4.2
numpy>=2.1.3
PyYAML>=6.0.2
scipy>=1.14.1
torch>=2.5.1
typing_extensions>=4.12.2
